requirements.txt
scikit-learn
sentence-transformers
streamlit
google-generativeai
